feat : Coupon Validation API 호출 적용 (#57)

+++ order/src/main/java/com/sparta/hotdeal/order/application/service/order/OrderService.java
@@ -1,24 +1,26 @@
 package com.sparta.hotdeal.order.application.service.order;
 
 import com.sparta.hotdeal.order.application.dtos.address.AddressDto;
-import com.sparta.hotdeal.order.application.dtos.coupon.CouponDto;
+import com.sparta.hotdeal.order.application.dtos.coupon.CouponValidationDto;
 import com.sparta.hotdeal.order.application.dtos.order.req.ReqPostOrderDto;
 import com.sparta.hotdeal.order.application.dtos.order.res.ResGetOrderByIdDto;
 import com.sparta.hotdeal.order.application.dtos.order.res.ResGetOrderListDto;
 import com.sparta.hotdeal.order.application.dtos.order_product.OrderProductDto;
 import com.sparta.hotdeal.order.application.dtos.product.ProductDto;
 import com.sparta.hotdeal.order.application.dtos.user.UserDto;
-import com.sparta.hotdeal.order.common.exception.ApplicationException;
-import com.sparta.hotdeal.order.common.exception.ErrorCode;
 import com.sparta.hotdeal.order.application.port.CouponClientPort;
 import com.sparta.hotdeal.order.application.port.ProductClientPort;
 import com.sparta.hotdeal.order.application.port.UserClientPort;
+import com.sparta.hotdeal.order.common.exception.ApplicationException;
+import com.sparta.hotdeal.order.common.exception.ErrorCode;
 import com.sparta.hotdeal.order.domain.entity.basket.Basket;
 import com.sparta.hotdeal.order.domain.entity.order.Order;
 import com.sparta.hotdeal.order.domain.repository.OrderRepository;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -38,7 +40,6 @@ public class OrderService {
     private final UserClientPort userClientPort;
 
     private final OrderProductService orderProductService;
-    private final OrderDetailService orderDetailService;
     private final OrderBasketService orderBasketService;
 
     public void createOrder(UUID userId, ReqPostOrderDto req) {
@@ -47,25 +48,28 @@ public class OrderService {
 
         //2. 상품 정보 조회
         List<UUID> productIds = basketList.stream().map(Basket::getProductId).toList();
-        Map<UUID, ProductDto> productMap = productClientPort.getProductAll(productIds);
+        List<ProductDto> productDtoList = productClientPort.getProductALL(productIds);
+        Map<UUID, ProductDto> productDtoMap = convertListToMap(productDtoList);
 
-        //3. 회사별 구매 상품 계산
-        Map<UUID, Integer> totalAmountByCompanyMap = orderDetailService.calculateTotalAmountByCompany(basketList,
-                productMap);
+        if (basketList.size() != productDtoList.size()) {
+            throw new ApplicationException(ErrorCode.ORDER_INVALID_VALUE_EXCEPTION);
+        }
 
         //4. 총 금액 계산
-        long totalAmount = orderDetailService.calculateTotalAmountFromCompanyMap(totalAmountByCompanyMap);
+        int totalAmount = calculateTotalAmount(basketList, productDtoMap);
 
         //5. 쿠폰 유효성 및 쿠폰 적용
+        int discountAmount = 0;
+        UUID couponId = null;
+        CouponValidationDto couponValidationDto = couponClientPort.validateCoupon(req.getCouponId(), basketList, productDtoMap);
 
-        CouponDto coupon = couponClientPort.getCoupon(req.getCouponId());
-        orderDetailService.validateCouponWithCompany(coupon, totalAmountByCompanyMap, totalAmount);
-        if (coupon != null) {
-            couponClientPort.useCoupon(coupon.getCouponId());
+        if(couponValidationDto.isValid()){
+            discountAmount = couponValidationDto.getTotalDiscountAmount();
+            couponId = req.getCouponId();
         }
 
         //6. 상품 감소 처리
-        productClientPort.reduceProductQuantity(basketList, productMap);
+//        productClientPort.reduceProductQuantity(basketList, productMap);
 
         //7. 장바구니 삭제
         orderBasketService.deleteBasketList(basketList);
@@ -76,12 +80,13 @@ public class OrderService {
                 address.getAddressId(),
                 userId,
                 totalAmount,
-                (coupon == null) ? null : coupon.getCouponId(),
-                (coupon == null) ? 0 : coupon.getDiscountAmount()
+                getOrderName(productDtoList),
+                couponId,
+                discountAmount
         );
 
         order = orderRepository.save(order);
-        orderProductService.createOrderProductList(order, basketList, productMap);
+        orderProductService.createOrderProductList(order, basketList, productDtoList);
         //결제 비동기 처리 추후 구현 필요
     }
 
@@ -135,4 +140,31 @@ public class OrderService {
         ));
     }
 
+    private int calculateTotalAmount(List<Basket> basketList, Map<UUID, ProductDto> productDtoMap) {
+        int totalAmount = 0;
+        for (Basket basket : basketList) {
+
+            ProductDto productDto = Optional.ofNullable(productDtoMap.get(basket.getProductId()))
+                    .orElseThrow(() -> new ApplicationException(ErrorCode.PRODUCT_NOT_FOUND_EXCEPTION));
+
+            if (productDto.getQuantity() < basket.getQuantity()) {
+                throw new ApplicationException(ErrorCode.PRODUCT_INVALID_QUANTITY_EXCEPTION);
+            }
+
+            int productPrice =
+                    (productDto.getDiscountPrice() == null) ? productDto.getPrice() : productDto.getDiscountPrice();
+            totalAmount += (productPrice * basket.getQuantity());
+        }
+        return totalAmount;
+    }
+
+    private String getOrderName(List<ProductDto> productDtoList) {
+        int listSize = productDtoList.size();
+        return productDtoList.get(0).getName() + "외 " + (listSize - 1) + "개";
+    }
+
+    private Map<UUID, ProductDto> convertListToMap(List<ProductDto> productDtoList) {
+        return productDtoList.stream()
+                .collect(Collectors.toMap(ProductDto::getProductId, productDto -> productDto));
+    }
 }
